config.guess: manual fixup after previous automated patches
This patch provides the special handling for the GNU system.  As these
were two small and unique edits, they were not included in the scripts.

This patch also cleans up other minor issues that must be addressed
before reverting to classic command substitutions and updates
"shellcheck" directives to account for changes in this script and the
change in "shellcheck" towards reporting individual portability issues.