config.guess: manual fixups after previous automatic patch
The tool could not handle command substitutions that span lines, but
fortunately there were only two such substitutions in the script.

The test for which universe is active on Pyramid is rewritten into a
case block because it was the only use of a command substitution as an
argument to the test command, which would require quoting.

* config.guess: Rewrite "if" for Pyramid systems to "case".