config.sub: testsuite: whitespace and sorting corrections.
Ensure that the second column of testsuite/config-sub.data is
consistently indented—always to the same distance, using tabs—
and that the entire file is properly sorted.  No functional change.

It might make sense to swap the two columns of this file and then
re-sort it, which would group all the aliases for a single canonical.
I think it would be easier to work with that way but I don’t want to
make that change unilaterally.

Signed-off-by: Dmitry V. Levin <ldv@altlinux.org>