config.sub: disable shellcheck warning SC2162 globally.
With the version of shellcheck I have (0.9.0), the existing attempts
to disable SC2162, above each use of read without -r, do not work
because they are not on the line *immediately* previous to the read
command — there’s always a “saved_IFS=$IFS” line in between.  Since
this script must not use read -r at all, we should just be disabling
SC2162 globally.

Signed-off-by: Dmitry V. Levin <ldv@altlinux.org>