config.sub: handle cNN-convex and convex-cNN shorthands correctly.
convex-cNN and cNN-convex (where NN in 1, 2, 32, 34, 38) are supposed
to be shorthands for cNN-convex-bsd, but the code to handle them
was in the wrong switch statement, making it unreachable.  Fix.

https://en.wikipedia.org/wiki/Convex_Computer says there was also a C4
machine in this product line, at least on paper, but it’s not clear to
me whether they ever shipped one.

Signed-off-by: Dmitry V. Levin <ldv@altlinux.org>