config.sub: don’t override OS for c90-cray.
Fix a typo in the logic for handling the OS default for c90-cray,
causing c90-cray-[anything] to get overridden to c90-cray-unicos.
I’m not aware of any non-Unicos operating systems for the C90, but
let’s not be a pointless obstacle for anyone who wants to write one.
(For testing purposes I used Linux, because config.sub’s generic logic
accepts it in combination with any CPU whatsoever.)

Signed-off-by: Dmitry V. Levin <ldv@altlinux.org>