config.sub: don’t override OS for blackfin, m68knommu, parisc.
A similar issue to the previous patch: make the OS default for the CPU
aliases “blackfin,” “m68knommu,” and “parisc” actually be a default
and not an override.  There have been many OSes for PA/RISC.  Both
blackfin and m68knommu are definitely used in bare-metal configuration
as well as with an OS.  And in general, CPU aliases should behave as
much like the canonical CPU name as possible.

Signed-off-by: Dmitry V. Levin <ldv@altlinux.org>