config.sub: improve consistency of shorthands and X-Y partial names.
It seems to me that if ‘fooco’ expands to ‘cpu-fooco-foonlyos’,
where ‘cpu’ is whatever CPU was hip at the time Fooco was making
computers, then ‘cpu-fooco’ should also expand to ‘cpu-fooco-foonlyos’,
and ‘cpu-foonlyos’ should as well, except when ‘foonlyos’ is something
generic like ‘sysv’ or ‘bsd’.  Implement a bunch of cases where we
didn’t do this quite right.  Probably there are more.

Signed-off-by: Dmitry V. Levin <ldv@altlinux.org>